+++ src/tools-TEST/BedAnnotateGC_Test.h
@@ -21,7 +21,7 @@ private slots:
         if (ref_file=="") SKIP("Test needs the reference genome!");
 
 		EXECUTE("BedAnnotateGC", "-clear -extend 20 -in " + TESTDATA("data_in/BedAnnotateGC_in2.bed") + " -out out/BedAnnotateGC_out2.bed -ref " + ref_file);
-		COMPARE_FILES_DELTA("out/BedAnnotateGC_out2.bed", TESTDATA("data_out/BedAnnotateGC_out2.bed"), 1); //delta because of macOS rounding problems
+		COMPARE_FILES_DELTA("out/BedAnnotateGC_out2.bed", TESTDATA("data_out/BedAnnotateGC_out2.bed"), 1.0, true, '\t'); //delta because of macOS rounding problems
 	}
 
 };